move helper files into stix_translation utils folder
