+++ stix_shifter/stix_translation/src/modules/aws_security_hub/aws_security_hub_translator.py
@@ -1,15 +1,11 @@
 from ..base.base_translator import BaseTranslator
 from ...json_to_stix.json_to_stix import JSONToStix
-# from stix_shifter.stix_translation.src.patterns.parser import generate_query
-# from . import data_mapping
 from . import query_constructor
 from os import path
-# from stix_shifter.stix_translation.src.unmapped_attribute_stripper import strip_unmapped_attributes
 
 
 class Translator(BaseTranslator):
     def transform_query(self, data, antlr_parsing_object, data_model_mapper, options, mapping=None):
-        # query_object = strip_unmapped_attributes(antlr_parsing_object, data_model_mapper)
 
         query_string = query_constructor.translate_pattern(
             antlr_parsing_object, data_model_mapper)
