+++ stix_shifter/stix_translation/src/modules/carbonblack/data_mapping.py
@@ -1,6 +1,6 @@
 from os import path
 import json
-from ...exceptions import DataMappingException
+from stix_shifter.stix_translation.src.utils.exceptions import DataMappingException
 
 
 class DataMapper:
