+++ stix_shifter/stix_translation/src/modules/csa/data_mapping.py
@@ -2,7 +2,7 @@ from os import path
 import json
 import re
 
-from stix_shifter.stix_translation.src.exceptions import DataMappingException
+from stix_shifter.stix_translation.src.utils.exceptions import DataMappingException
 
 
 def _fetch_mapping():
