+++ stix_shifter/stix_translation/src/modules/dummy/stix_to_query.py
@@ -1,10 +1,7 @@
 import logging
 
-# from ...patterns.parser import generate_query
 from ..base.base_query_translator import BaseQueryTranslator
-# from . import data_mapping
 from . import query_constructor
-# from stix_shifter.stix_translation.src.unmapped_attribute_stripper import strip_unmapped_attributes
 
 logger = logging.getLogger(__name__)
 
@@ -26,7 +23,6 @@ class StixToQuery(BaseQueryTranslator):
 
         logger.info("Converting STIX2 Pattern to data source query")
 
-        # query_object = strip_unmapped_attributes(antlr_parsing_object, data_model_mapper)
         query_string = query_constructor.translate_pattern(
             antlr_parsing_object, data_model_mapper, options)
         return query_string
