+++ stix_shifter/stix_translation/src/utils/exceptions.py
