+++ stix_shifter/stix_translation/src/utils/transformers.py
