+++ stix_shifter/stix_translation/src/utils/unmapped_attribute_stripper.py
@@ -1,6 +1,6 @@
-from .patterns.pattern_objects import StartStopQualifier, ObservationExpression, \
+from stix_shifter.stix_translation.src.patterns.pattern_objects import StartStopQualifier, ObservationExpression, \
     ComparisonExpression, Pattern, CombinedComparisonExpression, CombinedObservationExpression
-from stix_shifter.stix_translation.src.exceptions import DataMappingException
+from stix_shifter.stix_translation.src.utils.exceptions import DataMappingException
 
 
 class UnmappedAttributeStripper:
