+++ tests/csa_json_to_stix/test_class.py
@@ -1,5 +1,5 @@
 from stix_shifter.stix_translation.src.json_to_stix import json_to_stix_translator
-from stix_shifter.stix_translation.src import transformers
+from stix_shifter.stix_translation.src.utils import transformers
 from stix_shifter.stix_translation.src.modules.csa import csa_translator
 import json
 import unittest
