+++ tests/stix_translation/test_qradar_json_to_stix.py
@@ -1,5 +1,5 @@
 from stix_shifter.stix_translation.src.json_to_stix import json_to_stix_translator
-from stix_shifter.stix_translation.src import transformers
+from stix_shifter.stix_translation.src.utils import transformers
 from stix_shifter.stix_translation.src.modules.qradar import qradar_translator
 import json
 from stix_shifter.stix_translation import stix_translation
